using DataAccess in Location2IDQuadtree too, to make Android finally faster loading

+++ src/main/java/de/jetsli/graph/geohash/SpatialKeyAlgo.java
@@ -99,7 +99,7 @@ public class SpatialKeyAlgo implements KeyAlgo {
         setWorldBounds();
     }
 
-    public long getBits() {
+    public int getBits() {
         return allBits;
     }
